updated to 1.13 decompiled.

+++ src/main/java/com/eu/habbo/messages/outgoing/rooms/items/UpdateStackHeightTileHeightComposer.java
@@ -10,10 +10,10 @@ public class UpdateStackHeightTileHeightComposer extends MessageComposer
     private final HabboItem item;
     private final int height;
 
-    public UpdateStackHeightTileHeightComposer(HabboItem item, double height)
+    public UpdateStackHeightTileHeightComposer(HabboItem item, int height)
     {
         this.item = item;
-        this.height = (int)height;
+        this.height = height;
 
     }
